[Decomp][ObjectMgr] Cohesion split: extract graveyard loaders to ObjectMgrGraveyard.cpp
